+++ src/pages/index.ts
@@ -1,3 +1,4 @@
 export { default as Main } from './Main/Main';
-export { default as GardenEdit } from './Create/GardenEdit/GardenEdit';
-export { default as MyGardenEdit } from './Create/MyGardenEdit/MyGardenEdit';
+export { default as GardenEdit } from './Garden/GardenEdit/GardenEdit';
+export { default as MyGardenEdit } from './Garden/MyGardenEdit/MyGardenEdit';
+export { default as CommunityMain } from './Community/CommunityMain/CommunityMain';
